refactor: auth 여부에 따라 component 분기 후 api call
