+++ src/components/domain/TemplateHeader/UserNav/Authorized.jsx
@@ -4,16 +4,12 @@ import { Badge, Button, Icon } from '@/components/base';
 import AlarmModal from '@/components/domain/AlarmModal';
 import { useAuthContext } from '@/context/AuthProvider';
 import { COLOR } from '@/styles/color';
-import styled from '@emotion/styled';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-const UserNav = () => {
+const Authorized = () => {
   const navigate = useNavigate();
-  const {
-    authState: { isLoggedIn },
-    onLogout,
-  } = useAuthContext();
+  const { onLogout } = useAuthContext();
 
   const signOut = async () => {
     await postUserSignOut();
@@ -22,24 +18,10 @@ const UserNav = () => {
   };
 
   const [alarmModalVisible, setAlarmModalVisible] = useState(false);
-
-  if (!isLoggedIn) {
-    return (
-      <StyledHeaderUserNav>
-        <Button
-          style={{ width: '7.7rem', height: '2.1rem', margin: '0.7rem 0', padding: '0', fontSize: '1.8rem' }}
-          bgcolor={COLOR.HEADER_TRANSPARENT_BG}
-          onClick={() => navigate('/signIn')}>
-          로그인
-        </Button>
-      </StyledHeaderUserNav>
-    );
-  }
-
   const { data, isLoading, error } = useAlarms();
 
   return (
-    <StyledHeaderUserNav>
+    <>
       {alarmModalVisible && <AlarmModal visible={alarmModalVisible} onClose={() => setAlarmModalVisible(false)} />}
       <Icon name='users' size={2} onClick={() => navigate('/myGroup')} style={{ cursor: 'pointer' }} />
       <Badge dot={!isLoading && data && data.length && data.some(({ seen }) => !seen)}>
@@ -52,22 +34,8 @@ const UserNav = () => {
         onClick={() => signOut()}>
         로그아웃
       </Button>
-    </StyledHeaderUserNav>
+    </>
   );
 };
 
-export default UserNav;
-
-const StyledHeaderUserNav = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin-right: 2rem;
-  gap: 2rem;
-
-  position: relative;
-
-  & button:hover {
-    color: ${COLOR.GRAY3};
-  }
-`;
+export default Authorized;
